Astronomical Solutions ZB23.RXX for the past 3.6 Gyr

Astronomical Solutions ZB23.RXX for the past 3.6 Gyr

Format

get_solution("ZB23.Rxx")

A data frame with 8,750,001 rows and 5 columns:

time

Time in thousands of years (kyr).

ecc

Eccentricity \(e\) (unitless).

inc

Inclination \(I\) (radians).

obliquity

Obliqity \(\epsilon\) (radians).

cp

Climatic Precession \(e sin(\bar{\omega})\) (unitless).

Source

All astronomical solutions by Zeebe can be found on http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html.

They can be loaded into R quickly, using get_solution().

References

Zeebe, R. E. and Lourens, L. J. (2022). Geologically constrained astronomical solutions for the Cenozoic era. Earth and Planetary Science Letters. doi:10.1016/j.epsl.2022.117595