Full Astronomical Solution ZB18a for the past 100 Myr

The HNBody output of Zeebe & Lourens (2019) after some pre-processing using prepare_solution(). The wikipedia page on Orbital elements describes what the components relate to in order to uniquely specify an orbital plane. The function asks to download the files to the user's cache directory so that they can be accessed more quickly in the future.

Format

get_solution("full-ZB18a")

A data frame with 250,001 rows and 20 columns:

t

Time \(t\) (days).

time

Time in thousands of years (kyr).

aa

Semimajor axis \(a\) in astronomical units (au).

ee

Eccentricity \(e\) (unitless).

inc

Inclination \(I\) (degrees).

lph

Longitude of perihelion \(\varpi\) (degrees).

lan

Longitude of the ascending node \(\Omega\) (degrees).

arp

Argument of perihelion \(\omega\) (degrees).

mna

Mean anomaly \(M\) (degrees).

The following columns were computed from the above input with prepare_solution():
lphu

Unwrapped longitude of perihelion \(\varpi\) (degrees without jumps).

lanu

Unwrapped longitude of the ascending node \(\Omega\) (degrees without jumps).

hh

Variable: \(e\sin(\varpi)\).

kk

Variable: \(e\cos(\varpi)\).

pp

Variable: \(2\sin(0.5I)\sin(\Omega)\).

qq

Variable: \(2\sin(0.5I)\cos(\Omega)\).

cc

Helper: \(\cos(I)\).

dd

Helper: \(\cos(I)/2\).

nnx, nny, nnz

The \(x\), \(y\), and \(z\)-components of the Eart's orbit unit normal vector \(\vec{n}\), normal to Earth's instantaneous orbital plane.

Source

All astronomical solutions by Zeebe can be found on http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html.

They can be loaded into R quickly, using get_solution().

References

Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the Paleocene–Eocene boundary age constrained by geology and astronomy. Science, 365(6456), 926–929. doi:10.1126/science.aax0612 .

Zeebe, R. E. and Lourens, L. J. (2022). A deep-time dating tool for paleo-applications utilizing obliquity and precession cycles: The role of dynamical ellipticity and tidal dissipation. Paleoceanography and Paleoclimatology. doi:10.1029/2021PA004349

See also