Get an Astronomical Solution

Download supported astronomical solutions from the web and store it in the user's cache directory. The next use of the function will load the data from the cache rather than downloading it again. This also provides a wrapper for astrochron::getLaskar() if one of their supported solutions is specified, but converts the output to a tibble. Note that we do not cache these solutions locally, however.

Usage

get_solution(
  astronomical_solution = "full-ZB18a",
  quiet = FALSE,
  force = FALSE
)

Arguments

astronomical_solution

Character vector with the name of the desired solution. Defaults to "full-ZB18a".

quiet

Be quiet?

If TRUE, hide info messages.

If FALSE (the default) print info messages and timing.

force

Force re-downloading the results, even if the solution is saved to the cache.

Value

A tibble with the astronomical solution (and some preprocessed new columns).

References

Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the Paleocene–Eocene boundary age constrained by geology and astronomy. Science, 365(6456), 926–929. doi:10.1126/science.aax0612 .

Zeebe, R. E. and Lourens, L. J. (2022). A deep-time dating tool for paleo-applications utilizing obliquity and precession cycles: The role of dynamical ellipticity and tidal dissipation. Paleoceanography and Paleoclimatology. doi:10.1029/2021PA004349

See also

Examples

# \donttest{
get_solution("full-ZB18a")
#> ℹ The astronomical solution "full-ZB18a" has not been downloaded.
#> Would you like to download and process it now? (Yes/no/cancel) 
#> Would you like to save the astronomical solution to .csv and .rds? (Yes/no/cancel) 
#> ℹ Reading full-ZB18a.dat from website <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/PrecTilt/OS/ZB18a/ems-plan3.dat>.
#> ℹ Calculating helper columns.
#> Error in as.character(x): cannot coerce type 'closure' to vector of type 'character'
get_solution("ZB20a")
#> ℹ The astronomical solution "ZB20a" has not been downloaded.
#> Would you like to download and process it now? (Yes/no/cancel) 
#> Would you like to save the astronomical solution to .csv and .rds? (Yes/no/cancel) 
#> ℹ Reading ZB20a.dat from website <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/300Myr/ZB20a.dat>.
#> Error in as.character(x): cannot coerce type 'closure' to vector of type 'character'
get_solution("La11")
#> Warning: ℹ Relying on astrochron to get solution "La11"
#> ℹ We do not cache these results.
#> ! astrochron converts time from -kyr to ka by default.
#>  * Downloading Laskar et al. (2011) astronomical solution: La2011
#> 
#>    Please cite: Laskar, J., Fienga, A., Gastineau, M., Manche, H., 2011,
#>    La2010: A new orbital solution for the long-term motion of the Earth:
#>    Astron. Astrophys., Volume 532, A89.
#>   AND:
#>    Laskar, J., Gastineau, M., Delisle, J.-B., Farres, A., Fienga, A.: 2011,
#>    Strong chaos induced by close encounters with Ceres and Vesta:
#>    Astron. Astrophys., Volume 532, L4.
#>  * Decompressing solution
#> Warning: ℹ Output has column names "Time_ka" and "ecc_LA11"
#> # A tibble: 100,000 × 2
#>    Time_ka ecc_LA11
#>      <dbl>    <dbl>
#>  1       0   0.0167
#>  2       1   0.0172
#>  3       2   0.0175
#>  4       3   0.0178
#>  5       4   0.0182
#>  6       5   0.0185
#>  7       6   0.0187
#>  8       7   0.0189
#>  9       8   0.0192
#> 10       9   0.0193
#> # ℹ 99,990 more rows

# ignore the below, this is just to make CRAN builds happy!
# remove the directory with the cached astronomical solution to clean up
unlink(snvecR:::cachedir(), recursive = TRUE)
# }