Get an Astronomical Solution

Download supported astronomical solutions from the web and store it in the user's cache directory. The next use of the function will load the data from the cache rather than downloading it again. This also provides a wrapper for astrochron::getLaskar() if one of their supported solutions is specified, but converts the output to a tibble. Note that we do not cache these solutions locally, however.

Usage

get_solution(
  astronomical_solution = "full-ZB18a",
  quiet = FALSE,
  force = FALSE
)

Arguments

astronomical_solution

Character vector with the name of the desired solution. Defaults to "full-ZB18a".

quiet

Be quiet?

If TRUE, hide info messages.

If FALSE (the default) print info messages and timing.

force

Force re-downloading the results, even if the solution is saved to the cache.

Value

A tibble with the astronomical solution (and some preprocessed new columns).

References

Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the Paleocene–Eocene boundary age constrained by geology and astronomy. Science, 365(6456), 926–929. doi:10.1126/science.aax0612 .

Zeebe, R. E. and Lourens, L. J. (2022). A deep-time dating tool for paleo-applications utilizing obliquity and precession cycles: The role of dynamical ellipticity and tidal dissipation. Paleoceanography and Paleoclimatology. doi:10.1029/2021PA004349

See also

Examples

# \donttest{
get_solution("full-ZB18a")    # input for snvec
#> ℹ The astronomical solution "full-ZB18a" has not been cached.
#> ℹ Reading full-ZB18a.dat from website <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/PrecTilt/OS/ZB18a/ems-plan3.dat>.
#> ℹ Calculating helper columns.
#> ℹ The cache directory is /home/runner/.cache/R/snvecR.
#> ℹ Saved astronomical solution with helper columns full-ZB18a.rds to cache.
#> ℹ Future calls to `get_solution("full-ZB18a")` will read from the cache.
#> ! If you want to read from scratch, specify `force = TRUE`.
#> # A tibble: 250,001 × 20
#>           t  time    aa     ee   inc   lph   lan   arp   mna  lphu  lanu      hh
#>       <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>
#>  1        0   0    1.00 0.0167  7.15  27.3  180. -153. -2.45  27.3  180. 0.00767
#>  2  -146100  -0.4  1.00 0.0169  7.15  26.1 -180. -154.  1.27  26.1  180. 0.00742
#>  3  -292200  -0.8  1.00 0.0171  7.14  24.7 -179. -156.  5.22  24.7  181. 0.00713
#>  4  -438300  -1.2  1.00 0.0172  7.12  23.7 -179. -158.  8.75  23.7  181. 0.00690
#>  5  -584400  -1.6  1.00 0.0173  7.11  22.1 -178. -160. 12.8   22.1  182. 0.00653
#>  6  -730500  -2    1.00 0.0175  7.10  21.0 -178. -161. 16.4   21.0  182. 0.00627
#>  7  -876600  -2.4  1.00 0.0177  7.08  19.6 -178. -163. 20.3   19.6  182. 0.00594
#>  8 -1022700  -2.8  1.00 0.0178  7.06  18.5 -177. -164. 24.0   18.5  183. 0.00564
#>  9 -1168800  -3.2  1.00 0.0179  7.04  17.3 -177. -166. 27.7   17.3  183. 0.00531
#> 10 -1314900  -3.6  1.00 0.0181  7.02  15.8 -176. -168. 31.7   15.8  184. 0.00492
#> # ℹ 249,991 more rows
#> # ℹ 8 more variables: kk <dbl>, pp <dbl>, qq <dbl>, cc <dbl>, dd <dbl>,
#> #   nnx <dbl>, nny <dbl>, nnz <dbl>
get_solution("ZB18a-300")     # eccentricity
#> ℹ The astronomical solution "ZB18a-300" has not been cached.
#> ℹ Reading ZB18a-300.dat from website <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/300Myr/ZB18a.dat>.
#> ! Flipped time for "ZB18a-300" so that it is in negative kyr.
#> ℹ The cache directory is /home/runner/.cache/R/snvecR.
#> ℹ Saved astronomical solution with helper columns ZB18a-300.rds to cache.
#> ℹ Future calls to `get_solution("ZB18a-300")` will read from the cache.
#> ! If you want to read from scratch, specify `force = TRUE`.
#> # A tibble: 187,501 × 3
#>     time    ecc   inc
#>    <dbl>  <dbl> <dbl>
#>  1   0   0.0167  7.15
#>  2  -1.6 0.0173  7.11
#>  3  -3.2 0.0179  7.04
#>  4  -4.8 0.0184  6.95
#>  5  -6.4 0.0188  6.84
#>  6  -8   0.0192  6.70
#>  7  -9.6 0.0195  6.54
#>  8 -11.2 0.0196  6.36
#>  9 -12.8 0.0196  6.16
#> 10 -14.4 0.0196  5.94
#> # ℹ 187,491 more rows
get_solution("ZB20a")
#> ℹ The astronomical solution "ZB20a" has not been cached.
#> ℹ Reading ZB20a.dat from website <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/300Myr/ZB20a.dat>.
#> ! Flipped time for "ZB20a" so that it is in negative kyr.
#> ℹ The cache directory is /home/runner/.cache/R/snvecR.
#> ℹ Saved astronomical solution with helper columns ZB20a.rds to cache.
#> ℹ Future calls to `get_solution("ZB20a")` will read from the cache.
#> ! If you want to read from scratch, specify `force = TRUE`.
#> # A tibble: 187,501 × 3
#>     time    ecc   inc
#>    <dbl>  <dbl> <dbl>
#>  1   0   0.0167  7.15
#>  2  -1.6 0.0173  7.11
#>  3  -3.2 0.0179  7.04
#>  4  -4.8 0.0184  6.95
#>  5  -6.4 0.0188  6.84
#>  6  -8   0.0192  6.70
#>  7  -9.6 0.0195  6.54
#>  8 -11.2 0.0196  6.36
#>  9 -12.8 0.0196  6.16
#> 10 -14.4 0.0196  5.94
#> # ℹ 187,491 more rows
get_solution("La11")
#> Warning: ℹ Relying on astrochron to get solution "La11"
#> ℹ We do not cache these results.
#> ! astrochron converts time from -kyr to ka by default.
#>  * Downloading Laskar et al. (2011) astronomical solution: La2011
#> 
#>    Please cite: Laskar, J., Fienga, A., Gastineau, M., Manche, H., 2011,
#>    La2010: A new orbital solution for the long-term motion of the Earth:
#>    Astron. Astrophys., Volume 532, A89.
#>   AND:
#>    Laskar, J., Gastineau, M., Delisle, J.-B., Farres, A., Fienga, A.: 2011,
#>    Strong chaos induced by close encounters with Ceres and Vesta:
#>    Astron. Astrophys., Volume 532, L4.
#>  * Decompressing solution
#> Warning: ℹ Output has column names "Time_ka" and "ecc_LA11"
#> # A tibble: 100,000 × 2
#>    Time_ka ecc_LA11
#>      <dbl>    <dbl>
#>  1       0   0.0167
#>  2       1   0.0172
#>  3       2   0.0175
#>  4       3   0.0178
#>  5       4   0.0182
#>  6       5   0.0185
#>  7       6   0.0187
#>  8       7   0.0189
#>  9       8   0.0192
#> 10       9   0.0193
#> # ℹ 99,990 more rows
get_solution("PT-ZB18a(1,1)") # pre-computed precession-tilt
#> ℹ The astronomical solution "PT-ZB18a(1.0000,1.0000)" has not been cached.
#> ℹ Reading PT-ZB18a(1.0000,1.0000).dat from website <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/PrecTilt/ZB18a/asc/PT.De1.0000Td1.0000.dat>.
#> ℹ The cache directory is /home/runner/.cache/R/snvecR.
#> ℹ Saved astronomical solution with helper columns PT-ZB18a(1.0000,1.0000).rds
#>   to cache.
#> ℹ Future calls to `get_solution("PT-ZB18a(1.0000,1.0000)")` will read from the
#>   cache.
#> ! If you want to read from scratch, specify `force = TRUE`.
#> # A tibble: 249,480 × 4
#>      time   epl    phi     cp
#>     <dbl> <dbl>  <dbl>  <dbl>
#>  1  0     0.409 0      0.0163
#>  2 -0.379 0.410 0.0924 0.0167
#>  3 -0.760 0.411 0.185  0.0170
#>  4 -1.14  0.412 0.277  0.0171
#>  5 -1.52  0.413 0.369  0.0169
#>  6 -1.90  0.413 0.461  0.0165
#>  7 -2.28  0.414 0.552  0.0158
#>  8 -2.65  0.415 0.643  0.0150
#>  9 -3.07  0.416 0.743  0.0138
#> 10 -3.48  0.417 0.842  0.0124
#> # ℹ 249,470 more rows
get_solution("ZB23.R01")      # one of the 3.6 Gyr solutions
#> ℹ The astronomical solution "ZB23.R01" has not been cached.
#> ℹ Reading ZB23.R01.dat from website <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/3.5Gyr/ZB23-N64-eiop/ZB23.R01.eiop.dat.zip>.
#> Downloading any of the ZB23.RXX solutions will take some time.
#> Zip files are about 154 MB.
#> Continue downloading and caching? (Yes/no/cancel) 
#> ℹ The cache directory is /home/runner/.cache/R/snvecR.
#> ℹ Saved astronomical solution with helper columns ZB23.R01.rds to cache.
#> ℹ Future calls to `get_solution("ZB23.R01")` will read from the cache.
#> ! If you want to read from scratch, specify `force = TRUE`.
#> # A tibble: 8,750,001 × 5
#>     time    ecc   inc   epl     cp
#>    <dbl>  <dbl> <dbl> <dbl>  <dbl>
#>  1   0   0.0167 0.125 0.409 0.0163
#>  2  -0.4 0.0169 0.125 0.410 0.0168
#>  3  -0.8 0.0170 0.124 0.411 0.0170
#>  4  -1.2 0.0172 0.124 0.412 0.0170
#>  5  -1.6 0.0173 0.124 0.413 0.0168
#>  6  -2   0.0175 0.124 0.414 0.0163
#>  7  -2.4 0.0176 0.123 0.414 0.0155
#>  8  -2.8 0.0178 0.123 0.415 0.0146
#>  9  -3.2 0.0179 0.123 0.416 0.0134
#> 10  -3.6 0.0180 0.123 0.417 0.0120
#> # ℹ 8,749,991 more rows
# }