Get an Orbital Solution

Get an Orbital Solution

Usage

get_solution(orbital_solution = "ZB18a", quiet = FALSE, force = FALSE)

Arguments

orbital_solution

Character vector with the name of the orbital solution to use. One of "ZB18a" (default) from Zeebe and Lourens (2019), or "La11" (not yet implemented!).

quiet

Be quiet?

If TRUE, hide info messages.

If FALSE (the default) print info messages and timing.

force

Force re-downloading the results, even if the solution is saved to the cache.

Value

get_solution() returns a tibble with the orbital solution input and some preprocessed new columns.

References

Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the Paleocene–Eocene boundary age constrained by geology and astronomy. Science, 365(6456), 926–929. doi:10.1126/science.aax0612 .

See also

Examples

# \donttest{
get_solution()
#> # A tibble: 250,001 × 20
#>           t   age    aa     ee   inc   lph   lan   arp   mna  lphu  lanu      hh
#>       <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>
#>  1        0   0    1.00 0.0167  7.15  27.3  180. -153. -2.45  27.3  180. 0.00767
#>  2  -146100   0.4  1.00 0.0169  7.15  26.1 -180. -154.  1.27  26.1  180. 0.00742
#>  3  -292200   0.8  1.00 0.0171  7.14  24.7 -179. -156.  5.22  24.7  181. 0.00713
#>  4  -438300   1.2  1.00 0.0172  7.12  23.7 -179. -158.  8.75  23.7  181. 0.00690
#>  5  -584400   1.6  1.00 0.0173  7.11  22.1 -178. -160. 12.8   22.1  182. 0.00653
#>  6  -730500   2    1.00 0.0175  7.10  21.0 -178. -161. 16.4   21.0  182. 0.00627
#>  7  -876600   2.4  1.00 0.0177  7.08  19.6 -178. -163. 20.3   19.6  182. 0.00594
#>  8 -1022700   2.8  1.00 0.0178  7.06  18.5 -177. -164. 24.0   18.5  183. 0.00564
#>  9 -1168800   3.2  1.00 0.0179  7.04  17.3 -177. -166. 27.7   17.3  183. 0.00531
#> 10 -1314900   3.6  1.00 0.0181  7.02  15.8 -176. -168. 31.7   15.8  184. 0.00492
#> # ℹ 249,991 more rows
#> # ℹ 8 more variables: kk <dbl>, pp <dbl>, qq <dbl>, cc <dbl>, dd <dbl>,
#> #   nnx <dbl>, nny <dbl>, nnz <dbl>
# }