Get an Astronomical Solution

Get an Astronomical Solution

Usage

get_solution(astronomical_solution = "PT-ZB18a", quiet = FALSE, force = FALSE)

Arguments

astronomical_solution

Character vector with the name of the desired solution.

quiet

Be quiet?

If TRUE, hide info messages.

If FALSE (the default) print info messages and timing.

force

Force re-downloading the results, even if the solution is saved to the cache.

Value

A tibble with the astronomical solution (and some preprocessed new columns).

References

Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the Paleocene–Eocene boundary age constrained by geology and astronomy. Science, 365(6456), 926–929. doi:10.1126/science.aax0612 .

Zeebe, R. E. and Lourens, L. J. (2022). A deep-time dating tool for paleo-applications utilizing obliquity and precession cycles: The role of dynamical ellipticity and tidal dissipation. Paleoceanography and Paleoclimatology. doi:10.1029/2021PA004349

See also

Examples

# \donttest{
get_solution()
#> ℹ The astronomical solution PT-ZB18a has not been downloaded.
#> ℹ Reading PT-ZB18a.dat from website <http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/PrecTilt/OS/ZB18a/ems-plan3.dat>.
#> ℹ Calculating helper columns.
#> ℹ The cache directory is /home/runner/.cache/R/snvecR.
#> ℹ Saved PT-ZB18a.dat to cache.
#> ℹ Saved cleaned-up PT-ZB18a.csv to cache.
#> → Saved solution with helper columns PT-ZB18a.rds to cache.
#> # A tibble: 250,001 × 21
#>           t  t_ka   age    aa     ee   inc   lph   lan   arp   mna  lphu  lanu
#>       <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1        0   0     0    1.00 0.0167  7.15  27.3  180. -153. -2.45  27.3  180.
#>  2  -146100  -0.4   0.4  1.00 0.0169  7.15  26.1 -180. -154.  1.27  26.1  180.
#>  3  -292200  -0.8   0.8  1.00 0.0171  7.14  24.7 -179. -156.  5.22  24.7  181.
#>  4  -438300  -1.2   1.2  1.00 0.0172  7.12  23.7 -179. -158.  8.75  23.7  181.
#>  5  -584400  -1.6   1.6  1.00 0.0173  7.11  22.1 -178. -160. 12.8   22.1  182.
#>  6  -730500  -2     2    1.00 0.0175  7.10  21.0 -178. -161. 16.4   21.0  182.
#>  7  -876600  -2.4   2.4  1.00 0.0177  7.08  19.6 -178. -163. 20.3   19.6  182.
#>  8 -1022700  -2.8   2.8  1.00 0.0178  7.06  18.5 -177. -164. 24.0   18.5  183.
#>  9 -1168800  -3.2   3.2  1.00 0.0179  7.04  17.3 -177. -166. 27.7   17.3  183.
#> 10 -1314900  -3.6   3.6  1.00 0.0181  7.02  15.8 -176. -168. 31.7   15.8  184.
#> # ℹ 249,991 more rows
#> # ℹ 9 more variables: hh <dbl>, kk <dbl>, pp <dbl>, qq <dbl>, cc <dbl>,
#> #   dd <dbl>, nnx <dbl>, nny <dbl>, nnz <dbl>
# }